MAINT: Separate out pocketfft from the numpy additions
This commit ensures the pockefft.c file is identical to upstream.
